\name{titanic}
\alias{titanic}

\title{Passengers and Crew on the RMS Titanic}

\description{
  FIXME
}

\usage{data("titanic")}

\format{
A data frame containing 1405 observations on 28 variables.
\describe{
  \item{Survived}{.}
  \item{Pclass}{.}
  \item{Sex}{.}
  \item{Age}{.}
  \item{SibSp}{.}
  \item{Parch}{.}
  \item{Fare}{.}
  \item{Embarked}{.}
}
}

\details{
  FIXME
}

\source{
  FIXME
}

\references{
  FIXME
}

\examples{
data("titanic", package = "stablelearner")
summary(titanic)
}

\keyword{datasets}
